feat: 实现爬虫公开/私有功能 — 数据库模型增加 is_public 字段、API 接口改为基于 id 操作、新增可见性切换端点、前端支持公开/私有切换

+++ openspider/models/item.py
@@ -2,8 +2,7 @@
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Integer, String, func
-from sqlalchemy.dialects.mysql import JSON
+from sqlalchemy import DateTime, Integer, JSON, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from openspider.storage.database import Base
@@ -13,9 +12,9 @@ class ItemModel(Base):
     """爬取数据表"""
     __tablename__ = "items"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     spider_name: Mapped[str] = mapped_column(String(128), nullable=False, index=True)
-    task_id: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
+    task_id: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
     data: Mapped[dict] = mapped_column(JSON, nullable=False)
     url: Mapped[str] = mapped_column(String(2048), default="")
     crawled_at: Mapped[datetime] = mapped_column(
